requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
prompt-toolkit==3.0.3
python-dateutil==2.8.1
questionary==1.5.1
requests==2.22.0
six==1.14.0
toml==0.10.0
urllib3==1.25.8
wcwidth==0.1.8
